# openapi/v1.yaml
openapi: 3.1.0
info:
  title: Kitora REST API
  version: '1.0.0'
  summary: Public REST API for the Kitora SaaS platform.
  description: |
    Kitora's public API uses Bearer token authentication and returns JSON. Every
    request is rate-limited per token; expect `X-RateLimit-Remaining` and
    `X-RateLimit-Reset` (epoch seconds) on every response.

    ### Rate limits

    The default per-token bucket is **100 requests / 60 seconds** for all
    `/api/v1/*` routes, sliding-window. The bucket key is the token id, so
    horizontally-scaled callers should not race each other.

    On a 429 response, `X-RateLimit-Reset` is the epoch second at which the
    bucket refills — sleep until then and retry. Don't fan out new tokens to
    bypass the cap; tokens issued from the same org share a soft fair-use
    aggregate cap that triggers a temporary throttle.

    The `/api/openapi/v1.yaml` and `/docs/api` paths are public and not
    rate-limited (they're served from CDN with `Cache-Control: max-age=3600`).

    ### Webhook auto-disable

    Outbound webhook endpoints are **automatically paused** after 8
    consecutive failed deliveries (~44 hours of attempts under the default
    retry curve). When this happens:

    1. The endpoint's `disabledAt` is set to the time of the failure that
       crossed the threshold.
    2. Pending / retrying deliveries to that endpoint are mass-cancelled on
       the next cron tick.
    3. OWNER + ADMIN of the org receive an email.
...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/orgs/{slug}/identity-providers/{id}:
    parameters:
      - $ref: '#/components/parameters/OrgSlug'
      - $ref: '#/components/parameters/IdentityProviderId'
    get:
      tags: [SSO]
      summary: Get a single identity provider
      operationId: getIdentityProvider
      description: |
        Same shape as the list endpoint plus `samlMetadata` (the full XML).
      responses:
        '200':
          description: Provider detail
          headers:
            X-RateLimit-Remaining: { $ref: '#/components/headers/XRateLimitRemaining' }
            X-RateLimit-Reset: { $ref: '#/components/headers/XRateLimitReset' }
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    allOf:
                      - $ref: '#/components/schemas/IdentityProvider'
                      - type: object
                        properties:
                          samlMetadata:
                            type: [string, 'null']
                            description: Full SAML metadata XML, only set on SAML rows.
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
    patch:
      tags: [SSO]
      summary: Update an identity provider
      operationId: patchIdentityProvider
      description: |
        Partial update. Mutable fields:

        - `name`, `emailDomains`, `defaultRole`, `scimEnabled`
        - `enabledAt` (ISO timestamp to enable, `null` to flip back to draft)
        - `enforceForLogin` — OWNER only
        - `samlMetadata` (SAML rows only) / `oidcIssuer` / `oidcClientId`
        - `oidcClientSecret` — write-only; pass to rotate the encrypted secret
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IdentityProviderPatch' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean, const: true }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
    delete:
      tags: [SSO]
      summary: Delete an identity provider
      operationId: deleteIdentityProvider
      description: |
        Removes the IdP. Refused with `409 enforce-still-on` while
        `enforceForLogin` is true — turn enforcement off first to avoid
        accidentally locking the org out.
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean, const: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Conflict — `enforceForLogin` is still true.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
              example: { error: enforce-still-on }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: kitora_
      description: |
        Personal API tokens issued from `/settings/api-tokens`. Token format:
        `kitora_<base64url(32)>`. Tokens are scoped to a single organization
        and inherit the issuing user's role at issue-time.
  parameters:
    OrgSlug:
      in: path
      name: slug
      required: true
      schema: { type: string }
      description: Organization slug (URL-safe, lowercase). Must match the token's bound org.
      example: acme
...
    IdentityProviderId:
      in: path
      name: id
      required: true
      schema: { type: string }
      description: Identity provider id (cuid).
      example: cmrz1k1b00001zz000002zzz
  headers:
    XRateLimitRemaining:
      schema: { type: integer, format: int32, minimum: 0 }
      description: Tokens remaining in the current window.
      example: 99
    XRateLimitReset:
      schema: { type: integer, format: int64 }
      description: Epoch seconds when the bucket refills.
      example: 1745723404
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            invalidJson:
              summary: Malformed JSON body
              value: { error: invalid-json }
            invalidUrl:
              summary: URL failed SSRF check
              value: { error: invalid-url }
            unknownEvent:
              summary: Event not in registry
              value: { error: unknown-event, event: invoice.paid }
    Unauthorized:
      description: Missing, invalid, expired, or revoked token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: unauthorized }
    Forbidden:
      description: Authenticated but lacks org / role permission
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: forbidden }
    NotFound:
      description: Resource not found in this org
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: not-found }
    RateLimited:
      description: Token bucket empty
      headers:
        X-RateLimit-Remaining:
          schema: { type: integer, const: 0 }
        X-RateLimit-Reset: { $ref: '#/components/headers/XRateLimitReset' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: rate-limited }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Machine-readable error code. See the `info.description` table for the full list.
          example: unauthorized
        message:
          type: string
          description: Optional human-readable detail. Not always present.
        event:
          type: string
          description: Echoed back on `unknown-event` to identify which element failed.
...
    SsoProtocol:
      type: string
      description: SAML 2.0 vs OIDC. One of each per org max.
      enum: [SAML, OIDC]
    IdentityProvider:
      type: object
      required:
        - id
        - name
        - protocol
        - emailDomains
        - defaultRole
        - enforceForLogin
        - enabledAt
        - scimEnabled
        - scimTokenPrefix
        - createdAt
        - updatedAt
      properties:
        id: { type: string }
        name: { type: string, maxLength: 120, example: 'Okta production' }
        protocol: { $ref: '#/components/schemas/SsoProtocol' }
        emailDomains:
          type: array
          items: { type: string, example: 'acme.com' }
          description: Lowercase RFC 1123 hostnames. Used to route /login → this IdP.
        defaultRole:
          type: string
          enum: [OWNER, ADMIN, MEMBER]
          description: Role assigned to JIT-provisioned members.
        enforceForLogin:
          type: boolean
          description: When true, the org's password / OAuth login paths are blocked.
        enabledAt:
          oneOf:
            - { type: string, format: date-time }
            - type: 'null'
          description: Set to a timestamp to make this IdP available on `/login`. `null` = draft.
        scimEnabled: { type: boolean }
        scimTokenPrefix:
          oneOf:
            - { type: string, maxLength: 8 }
            - type: 'null'
          description: First 8 chars of the SCIM token body, safe to display.
        oidcIssuer:
          oneOf:
            - { type: string, format: uri }
            - type: 'null'
          description: Set on OIDC rows.
        oidcClientId:
          oneOf:
            - { type: string }
            - type: 'null'
          description: Set on OIDC rows. Client secret is write-only.
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    IdentityProviderPatch:
      type: object
      description: All fields optional — pass any subset.
      properties:
        name: { type: string, maxLength: 120 }
        emailDomains:
          type: array
          items: { type: string }
        defaultRole:
          type: string
          enum: [OWNER, ADMIN, MEMBER]
        enforceForLogin: { type: boolean, description: 'OWNER-only field.' }
        enabledAt:
          oneOf:
            - { type: string, format: date-time }
            - type: 'null'
        scimEnabled: { type: boolean }
        samlMetadata: { type: string }
        oidcIssuer: { type: string, format: uri }
        oidcClientId: { type: string }
        oidcClientSecret:
          type: string
          description: Pass to rotate the encrypted secret. Write-only.